requirements.txt
httpx>=0.27,<1
paho-mqtt>=2,<3
